--- app/routes/ffxiv.tsx
@@ -1,9 +1,0 @@
-import { Outlet } from '@remix-run/react'
-
-export default function FFXIV() {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  )
-}
