--- app/routes/why.tsx
@@ -1,9 +1,0 @@
-import { Outlet } from '@remix-run/react'
-
-export default function Why() {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  )
-}
